+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: ability to pass custom from generators to ruleSet is not possible
 * 🎉 feat: added AlwaysCheckRules method
 * 🎉 feat: added WhenNotExistOne + WhenNotExistAll